+++ src/routes/pages/Login.css
@@ -24,19 +24,19 @@ margin-top: 240px;
   color: #4D4F50;
   
 }
-.email-login{
+button.email-login{
   height: 52px;
   width: calc(100%-4px);
   background-color: white;
   border: 2px solid #1E1F1F;
-  align-items: center;  
+  align-items: center;
   justify-content: center;
   display: flex;
   border-radius: 5px;
   font-size: 18px;
   margin-top: 16px;
   color: #1E1F1F;
-  
+
 }
 .kakao-login{
